感谢LAzysheep小姐姐，签到懒人版
